added first error test case
